Added unit test for RDPDR channel manager.

+++ src/utils/virtual_channel_data_sender.hpp
@@ -26,8 +26,8 @@ class VirtualChannelDataSender
 public:
     virtual ~VirtualChannelDataSender() = default;
 
-    virtual void operator()(size_t total_length, uint32_t flags,
-        const uint8_t * chunk_data, size_t chunk_data_length) = 0;
+    virtual void operator()(uint32_t total_length, uint32_t flags,
+        const uint8_t * chunk_data, uint32_t chunk_data_length) = 0;
 };
 
 #endif // #ifndef _REDEMPTION_UTILS_VIRTUAL_CHANNEL_DATA_SENDER_HPP_
